requirements.txt
openai
fastapi
black
isort
websockets
pydantic
langchain==0.0.194
uvicorn
jinja2
faiss-cpu
unstructured
libmagic
beautifulsoup4
tiktoken
python-dotenv

# NEEDED FOR CHROMADB
sentence_transformers
chromadb==0.3.25